✨ Added Explore by topic to the Network tab
