+++ apps/activitypub/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tryghost/activitypub",
-  "version": "1.0.37",
+  "version": "1.1.0",
   "license": "MIT",
   "repository": {
     "type": "git",
